+++ src/services/api.js
@@ -1,19 +1,30 @@
 import axios from 'axios';
 import {API_URL} from '@env';
 import {getUserIdToken} from '@app/services';
-import {storeData} from '@app/util';
 
 const createUserInDb = async (email) => {
   try {
     const idToken = await getUserIdToken();
-    const res = await axios.post(
+    await axios.post(
       `${API_URL}/pushie/user`,
       {email},
       {headers: {Authorization: `Bearer ${idToken}`}},
     );
-    await storeData('apiKey', res.data);
   } catch (err) {
-    console.error(err);
+    throw new Error(err.message);
+  }
+};
+
+const deleteNotification = async (item) => {
+  try {
+    const idToken = await getUserIdToken();
+    const res = await axios.delete(`${API_URL}/pushie/user`, {
+      headers: {Authorization: `Bearer ${idToken}`},
+      data: item,
+    });
+    return res;
+  } catch (err) {
+    throw new Error(err.message);
   }
 };
 
@@ -25,7 +36,7 @@ const getNotifications = async () => {
     });
     return res.data;
   } catch (err) {
-    console.error(err);
+    throw new Error(err.message);
   }
 };
 
@@ -37,7 +48,7 @@ const getUserApiKey = async () => {
     });
     return res.data;
   } catch (err) {
-    console.error(err);
+    throw new Error(err.message);
   }
 };
 
@@ -50,8 +61,14 @@ const saveUserDeviceToken = async (token) => {
       {headers: {Authorization: `Bearer ${idToken}`}},
     );
   } catch (err) {
-    console.error(err);
+    throw new Error(err.message);
   }
 };
 
-export {createUserInDb, getNotifications, getUserApiKey, saveUserDeviceToken};
+export {
+  createUserInDb,
+  deleteNotification,
+  getNotifications,
+  getUserApiKey,
+  saveUserDeviceToken,
+};