+++ src/util/index.js
@@ -1,4 +1,19 @@
 import {getLocalDateTime} from './dateTime';
+import {
+  getDeviceBiometrics,
+  storeInKeychain,
+  getFromKeychain,
+  removeFromKeychain,
+} from './keychain';
 import {storeData, getData, clearStorage} from './storage';
 
-export {getLocalDateTime, storeData, getData, clearStorage};
+export {
+  getLocalDateTime,
+  getDeviceBiometrics,
+  storeInKeychain,
+  getFromKeychain,
+  removeFromKeychain,
+  storeData,
+  getData,
+  clearStorage,
+};