+++ src/views/Home.js
@@ -6,12 +6,20 @@ import {
   FlatList,
   StyleSheet,
   View,
+  Share,
 } from 'react-native';
 import messaging from '@react-native-firebase/messaging';
 import Icon from 'react-native-vector-icons/MaterialIcons';
 import Spinner from 'react-native-spinkit';
+import ContextMenu from 'react-native-context-menu-view';
 import {NotificationRow, NotificationModal} from '@app/components';
-import {getNotifications} from '@app/services';
+import {
+  deleteNotification,
+  getNotifications,
+  getUserApiKey,
+} from '@app/services';
+import {storeInKeychain} from '@app/util';
+import {CONTEXT_DELETE, CONTEXT_PREVIEW, CONTEXT_SHARE} from '@app/constants';
 
 const Home = ({navigation}) => {
   const [modalVisible, setModalVisible] = useState(false);
@@ -24,25 +32,87 @@ const Home = ({navigation}) => {
       loadNotifications();
     });
 
+    getUserApiKey()
+      .then(async (value) => {
+        await storeInKeychain('api', value);
+      })
+      .catch((err) => {
+        console.error(err);
+      });
+
     return unsubscribe;
   }, []);
 
+  const removeNotification = async (item) => {
+    deleteNotification(item)
+      .then(() => {
+        setNotifications(notifications.filter((o) => o.id !== item.id));
+      })
+      .catch((err) => {
+        console.error(err);
+      });
+  };
+
+  const onShare = async (item) => {
+    try {
+      await Share.share(
+        {
+          message: `${item.title}\n${
+            item.description ? item.description : item.shortDescription
+          }`,
+        },
+        {
+          subject: item.title,
+          tintColor: '#0080FF',
+        },
+      );
+    } catch (err) {
+      console.error(err);
+    }
+  };
+
   const renderNotificationRow = ({item}) => (
-    <NotificationRow
-      title={item.title}
-      shortDescription={item.shortDescription}
-      timestamp={item.timestamp}
-      onPress={() => {
-        setNotificationItem(item);
-        setModalVisible(true);
-      }}
-    />
+    <ContextMenu
+      previewBackgroundColor={'white'}
+      actions={[
+        {
+          title: CONTEXT_SHARE,
+          systemIcon: 'square.and.arrow.up',
+          destructive: false,
+        },
+        {title: CONTEXT_DELETE, systemIcon: 'trash', destructive: true},
+      ]}
+      onPress={(event) => {
+        const {name} = event.nativeEvent;
+        if (name === CONTEXT_DELETE) {
+          removeNotification(item);
+        } else if (name === CONTEXT_PREVIEW) {
+          setNotificationItem(item);
+          setModalVisible(true);
+        } else if (name === CONTEXT_SHARE) {
+          onShare(item);
+        }
+      }}>
+      <NotificationRow
+        title={item.title}
+        shortDescription={item.shortDescription}
+        timestamp={item.timestamp}
+        onPress={() => {
+          setNotificationItem(item);
+          setModalVisible(true);
+        }}
+      />
+    </ContextMenu>
   );
 
   const loadNotifications = async () => {
     setLoading(true);
-    const res = await getNotifications();
-    setNotifications(res);
+    try {
+      const res = await getNotifications();
+      setNotifications(res);
+    } catch (err) {
+      console.error(err);
+    }
     setLoading(false);
   };
 
@@ -52,7 +122,7 @@ const Home = ({navigation}) => {
 
   return (
     <>
-      <StatusBar barStyle={'dark-content'} />
+      <StatusBar barStyle={'dark-content'} animated={true} />
       <NotificationModal
         visible={modalVisible}
         item={notificationItem}
@@ -69,7 +139,7 @@ const Home = ({navigation}) => {
               name={'settings'}
               size={30}
               style={styles.settingsButton}
-              onPress={() => navigation.navigate('Settings')}
+              onPress={() => navigation.navigate('SettingsStack')}
             />
           </View>
         </View>
@@ -83,7 +153,7 @@ const Home = ({navigation}) => {
             onRefresh={() => loadNotifications()}
             data={notifications}
             renderItem={renderNotificationRow}
-            keyExtractor={(notificationItem) => notificationItem.id}
+            keyExtractor={(item) => item.id}
           />
         )}
       </SafeAreaView>