+++ src/views/index.js
@@ -1,6 +1,10 @@
+import About from './About';
+import Api from './Api';
+import Changelog from './Changelog';
+import Help from './Help';
 import Home from './Home';
+import License from './License';
 import Login from './Login';
 import Settings from './Settings';
-import SplashScreen from './SplashScreen';
 
-export {Home, Login, Settings, SplashScreen};
+export {About, Api, Changelog, Help, Home, License, Login, Settings};